backend/pyproject.toml
[project]
name = "personal-book-brain"
version = "0.1.0"
description = "Personal Book Brain"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "defusedxml>=0.7.1",
    "fastapi[standard]>=0.128.0",
    "firebase-admin>=7.1.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-cloud-discoveryengine>=0.16.0",
    "google-cloud-firestore>=2.22.0",
    "google-genai>=1.57.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",  # one-blank-line-before-class (incompatible with D211)
    "D213",  # multi-line-summary-second-line (incompatible with D212)
    "COM812", # missing-trailing-comma (formatter conflict)
    "ISC001", # single-line-implicit-string-concatenation (formatter conflict)
    "E501",   # line-too-long
]
allowed-confusables = ["（", "）", "【", "】", "〜"]
